move predict.py to serve folder

+++ serve/predict.py
@@ -89,6 +89,6 @@ def predict_fn(input_data, model):
     with torch.no_grad():
         output = model.forward(data)
 
-    result = np.round(output.numpy())
+    result = int(np.round(output.numpy()))
 
     return result
